CHange the main loop to displayinvalidate rect
Update and move inter_rect to return a size zero rect if there is no
intersection.
Add draw() function to draw all the widget for all the rectangle of the
list.
Change & fix invalidate rect.
Uncomment geomnotify call in runfunc.
Add calls to invalidate rect when a widget is configured, placed